pyproject.toml
[project]
name = "profile-comparison"
version = "0.1.0"
description = ""
authors = [
    {name = "Kunal7069",email = "jaink7069@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "openai (>=1.75.0,<2.0.0)",
    "tokencost (>=0.1.20,<0.2.0)"
]

[tool.poetry]
packages = [{include = "profile_comparison", from = "src"}]


[tool.poetry.group.dev.dependencies]
uvicorn = {extras = ["standard"], version = "^0.34.1"}

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start-api = "test:app"